+++ CotEditor/Sources/SyntaxParser.swift
@@ -196,6 +196,11 @@ extension SyntaxParser {
         guard UserDefaults.standard[.enableSyntaxHighlight] else { return nil }
         guard !self.textStorage.string.isEmpty else { return nil }
         
+        guard !self.style.isNone else {
+            self.textStorage.apply(highlights: [:], range: self.textStorage.range)
+            return nil
+        }
+        
         let wholeRange = self.textStorage.range
         
         // use cache if the content of the whole document is the same as the last
@@ -280,7 +285,7 @@ extension SyntaxParser {
         
         assert(Thread.isMainThread)
         
-        guard !highlightRange.isEmpty else { return nil }
+        guard !highlightRange.isEmpty, !self.style.isNone else { return nil }
         
         // just clear current highlight and return if no coloring needs
         guard self.style.hasHighlightDefinition else {