+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Change Log
 
 ### Fixes
 
+- [beta] Fix an issue where the current syntax highlight was not removed when selecting style “None.”
 - [beta] Fix an issue where the visibility of invisibles of opened documents cannot be enabled.